# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Gerenciamento Clinico",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/exam-results": {
            "get": {
                "tags": [
                    "Resultados de Exames"
                ],
                "summary": "Listar resultados de exames",
                "description": "Retorna uma lista de resultados de exames",
                "operationId": "cef213570cab27d189a06ca0f8a7f9f0",
                "responses": {
                    "200": {
                        "description": "Lista de resultados de exames",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ExamResult"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Resultados de Exames"
                ],
                "summary": "Criar resultado de exame",
                "description": "Cria um novo resultado de exame",
                "operationId": "124e4e53544154f686050d75262c4658",
                "requestBody": {
                    "$ref": "#/components/requestBodies/StoreExamResultRequest"
                },
                "responses": {
                    "201": {
                        "description": "Resultado de exame criado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ExamResult"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/exam-results/{id}": {
            "get": {
                "tags": [
                    "Resultados de Exames"
                ],
                "summary": "Exibir resultado de exame",
                "description": "Retorna os dados de um resultado de exame específico",
                "operationId": "311415169521e5cb852b46f4f7473355",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do resultado de exame",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dados do resultado de exame",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ExamResult"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Resultados de Exames"
                ],
                "summary": "Atualizar resultado de exame",
                "description": "Atualiza os dados de um resultado de exame específico",
                "operationId": "db835c96a26e40f7ca54009a5b48bf0b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do resultado de exame",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "$ref": "#/components/requestBodies/UpdateExamResultRequest"
                },
                "responses": {
                    "200": {
                        "description": "Resultado de exame atualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ExamResult"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Resultados de Exames"
                ],
                "summary": "Excluir resultado de exame",
                "description": "Exclui um resultado de exame específico",
                "operationId": "a77a0970d9f672530b4e6fd4ea8fe5c1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do resultado de exame",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Resultado de exame excluído"
                    }
                }
            }
        },
        "/api/exam-schedulings": {
            "get": {
                "tags": [
                    "Agendamentos de Exames"
                ],
                "summary": "Listar agendamentos de exames",
                "description": "Retorna uma lista de agendamentos de exames",
                "operationId": "df6ebf1dc9d4799b181a167ebfb31e9a",
                "responses": {
                    "200": {
                        "description": "Lista de agendamentos de exames",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ExamScheduling"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Agendamentos de Exames"
                ],
                "summary": "Criar agendamento de exame",
                "description": "Cria um novo agendamento de exame",
                "operationId": "f481ec5a9a9039834c804c3835fba499",
                "requestBody": {
                    "$ref": "#/components/requestBodies/StoreExamSchedulingRequest"
                },
                "responses": {
                    "201": {
                        "description": "Agendamento de exame criado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ExamScheduling"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/exam-schedulings/{id}": {
            "get": {
                "tags": [
                    "Agendamentos de Exames"
                ],
                "summary": "Exibir agendamento de exame",
                "description": "Retorna os dados de um agendamento de exame específico",
                "operationId": "3a021b53473a9476091f4cf8f5b56b47",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do agendamento de exame",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dados do agendamento de exame",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ExamScheduling"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Agendamentos de Exames"
                ],
                "summary": "Atualizar agendamento de exame",
                "description": "Atualiza os dados de um agendamento de exame específico",
                "operationId": "fe55117daaef69abcbf203f938faabf9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do agendamento de exame",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "$ref": "#/components/requestBodies/UpdateExamSchedulingRequest"
                },
                "responses": {
                    "200": {
                        "description": "Agendamento de exame atualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ExamScheduling"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Agendamentos de Exames"
                ],
                "summary": "Excluir agendamento de exame",
                "description": "Exclui um agendamento de exame específico",
                "operationId": "9f368313e3b88c4f2c35c62043bee3e2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do agendamento de exame",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Agendamento de exame excluído"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ExamResult": {
                "title": "Resultado de Exame",
                "description": "Modelo de Resultado de Exame",
                "required": [
                    "exam_scheduling_id",
                    "result",
                    "result_date"
                ],
                "properties": {
                    "id": {
                        "description": "ID do resultado do exame",
                        "type": "integer"
                    },
                    "exam_scheduling_id": {
                        "description": "ID do agendamento do exame",
                        "type": "integer"
                    },
                    "result": {
                        "description": "Resultado do exame",
                        "type": "string"
                    },
                    "observations": {
                        "description": "Observações do exame",
                        "type": "string"
                    },
                    "result_date": {
                        "description": "Data do resultado do exame",
                        "type": "string",
                        "format": "date"
                    },
                    "created_at": {
                        "description": "Data de criação",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Data de atualização",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "ExamScheduling": {
                "title": "Agendamento de Exame",
                "description": "Modelo de Agendamento de Exame",
                "required": [
                    "client_id",
                    "medic_id",
                    "exam_id",
                    "scheduled_date",
                    "scheduled_time"
                ],
                "properties": {
                    "id": {
                        "description": "ID do agendamento do exame",
                        "type": "integer"
                    },
                    "client_id": {
                        "description": "ID do cliente",
                        "type": "integer"
                    },
                    "medic_id": {
                        "description": "ID do médico",
                        "type": "integer"
                    },
                    "exam_id": {
                        "description": "ID do exame",
                        "type": "integer"
                    },
                    "scheduled_date": {
                        "description": "Data do agendamento",
                        "type": "string",
                        "format": "date"
                    },
                    "scheduled_time": {
                        "description": "Hora do agendamento",
                        "type": "string",
                        "format": "time"
                    },
                    "times_rescheduled": {
                        "description": "Número de vezes reagendado",
                        "type": "integer"
                    },
                    "status": {
                        "description": "Status do agendamento",
                        "type": "string"
                    },
                    "created_at": {
                        "description": "Data de criação",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Data de atualização",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "requestBodies": {
...
            "StoreExamResultRequest": {
                "description": "Dados para criar um resultado de exame",
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "exam_scheduling_id": {
                                    "description": "ID do agendamento do exame",
                                    "type": "integer"
                                },
                                "result": {
                                    "description": "Resultado do exame",
                                    "type": "string"
                                },
                                "observations": {
                                    "description": "Observações do exame",
                                    "type": "string"
                                },
                                "result_date": {
                                    "description": "Data do resultado do exame",
                                    "type": "string",
                                    "format": "date"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "StoreExamSchedulingRequest": {
                "description": "Dados para criar um agendamento de exame",
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "client_id": {
                                    "description": "ID do cliente",
                                    "type": "integer"
                                },
                                "medic_id": {
                                    "description": "ID do médico",
                                    "type": "integer"
                                },
                                "exam_id": {
                                    "description": "ID do exame",
                                    "type": "integer"
                                },
                                "scheduled_date": {
                                    "description": "Data do agendamento",
                                    "type": "string",
                                    "format": "date"
                                },
                                "scheduled_time": {
                                    "description": "Hora do agendamento",
                                    "type": "string",
                                    "format": "time"
                                },
                                "status": {
                                    "description": "Status do agendamento",
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
...
            "UpdateExamResultRequest": {
                "description": "Dados para atualizar um resultado de exame",
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "exam_scheduling_id": {
                                    "description": "ID do agendamento do exame",
                                    "type": "integer"
                                },
                                "result": {
                                    "description": "Resultado do exame",
                                    "type": "string"
                                },
                                "observations": {
                                    "description": "Observações do exame",
                                    "type": "string"
                                },
                                "result_date": {
                                    "description": "Data do resultado do exame",
                                    "type": "string",
                                    "format": "date"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "UpdateExamSchedulingRequest": {
                "description": "Dados para atualizar um agendamento de exame",
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "client_id": {
                                    "description": "ID do cliente",
                                    "type": "integer"
                                },
                                "medic_id": {
                                    "description": "ID do médico",
                                    "type": "integer"
                                },
                                "exam_id": {
                                    "description": "ID do exame",
                                    "type": "integer"
                                },
                                "scheduled_date": {
                                    "description": "Data do agendamento",
                                    "type": "string",
                                    "format": "date"
                                },
                                "scheduled_time": {
                                    "description": "Hora do agendamento",
                                    "type": "string",
                                    "format": "time"
                                },
                                "times_rescheduled": {
                                    "description": "Número de vezes reagendado",
                                    "type": "integer"
                                },
                                "status": {
                                    "description": "Status do agendamento",
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}